Move FFE metric encoding into datadog-ffe
