+++ datadog-ffe/src/lib.rs
@@ -4,5 +4,7 @@
 mod flag_type;
 
 pub mod rules_based;
+#[cfg(feature = "telemetry")]
+pub mod telemetry;
 
 pub use flag_type::{ExpectedFlagType, FlagType};
