+++ datadog-ffe/src/telemetry/mod.rs
@@ -1,0 +1,13 @@
+// Copyright 2026-Present Datadog, Inc. https://www.datadoghq.com/
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod evaluation_metrics;
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Default, Deserialize, Eq, Hash, PartialEq, Serialize)]
+pub struct FfeTelemetryContext {
+    pub service: String,
+    pub env: String,
+    pub version: String,
+}
