+++ datadog-sidecar/Cargo.toml
@@ -29,6 +29,7 @@ libdd-trace-utils = { path = "../libdd-trace-utils" }
 libdd-trace-stats = { path = "../libdd-trace-stats", default-features=false, features = ["https"] }
 datadog-remote-config = { path = "../datadog-remote-config" , features = ["live-debugger"]}
 datadog-live-debugger = { path = "../datadog-live-debugger" }
+datadog-ffe = { path = "../datadog-ffe", features = ["telemetry"] }
 libdd-crashtracker = { path = "../libdd-crashtracker" }
 libdd-dogstatsd-client = { path = "../libdd-dogstatsd-client" }
 libdd-tinybytes = { path = "../libdd-tinybytes" }
@@ -43,12 +44,10 @@ datadog-ipc-macros = { path = "../datadog-ipc-macros" }
 
 rand = "0.8.3"
 rmp-serde = "1.3.0"
-libdd-trace-protobuf = { path = "../libdd-trace-protobuf" }
 serde = { version = "1.0", features = ["derive", "rc"] }
 serde_with = "3.6.0"
 bincode = { version = "1.3.3" }
 serde_json = "1.0"
-prost = "0.14.1"
 base64 = "0.22.1"
 spawn_worker = { path = "../spawn_worker" }
 zwohash = "0.1.2"
